contacts/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.4
psycopg2-binary==2.9.10
redis==5.2.0
alembic==1.14.1
SQLAlchemy==2.0.36

